refactor: voice chat join message
Rewrite the event handler in a deticated file voiceChatJoinMessage.js loaded dynamicly by index.js.

The events makes use now of the utility function pickRandomFile.
The refactor is more readable and takes into account more errors.
The refactor returns early at more points to save runtime.
