+++ index.js
@@ -115,25 +115,3 @@ function bdayHandler(message){
     // update last cheak date to today
     bday_cheak = today;
 }
-
-bot.on('voiceStateUpdate', (oldState, newState) => {
-
-  let oldChannel = oldState.channelId;
-  let newChannel = newState.channelId;
-
-  if(oldChannel === null && newChannel != null){
-    // user joined voice chat, wellcome them
-    let user = newState.member.user;  // find user who joined
-    if (user.bot === false) {  // dont triger from bots
-        Command.doPikaNoise(newState.channel);
-    }
-  } else if (oldChannel != null && newState === null) {
-    let user = oldState.member.user;  // find user who left
-    // user left voice chat :< goodbye
-  } else if (oldChannel != null && newChannel != null){
-    let user = newState.member.user;  // find user who moved
-    // user moved to a new voice channel or changed state
-  } else {
-    // user is a hacker
-  }
-})
